==> codes/23_extendedFig5c_2.R <==
# This code doesn't perform any analysis, it only generates the data for GWAS and RF.
# For Generalists.

library(caret)
library(dplyr)

# Numbers for random seeds
set.seed(32); numbers = sample(x = 1:100000, size = 1000, replace = FALSE)

# Generate partitions
partition = function(s) {
  set.seed(s)
  inTrain = createDataPartition(y = 1:212, p = 0.8, list = FALSE)
  return(inTrain)
}

parts = list()
for(n in 1:1000) {
  parts[[n]] = partition(numbers[n])
}

# Abundance data
pheno = read.delim('results/cumulativeAbundance.csv', sep = ',')
pheno = pheno[, c('Line1', 'raG')]
pheno = pheno %>% group_by(Line1) %>% summarise(Generalists = mean(raG))
pheno$Generalists = round(pheno$Generalists, 6)

# To GWAS format
template = read.delim('data/templateForGWAS.csv', sep = ',')
class(pheno) = class(pheno)[3]
row.names(pheno) = pheno$Line1
pheno = pheno[template$Taxa,]
row.names(pheno) = NULL
colnames(pheno)[1] = 'Taxa'
write.csv(pheno, 'results/generalistsForGWAS_CRA.csv')

# Write out the data tables
writer = function(x, i) {
  df = x[i, ]
  write.table(df,
              paste0('results/inputForGWAS_CRA/generalists/generalists', as.character(numbers[n]), '.csv'),
              quote = TRUE, sep = ',',
              row.names = FALSE, col.names = TRUE)
}

# Only first 100
for(n in 1:100) {
  writer(pheno, parts[[n]])
}
